refactor: move content types into own files
